requirements.txt
streamlit>=1.24.0
openpyxl>=3.0.10
pandas>=1.5.0
python-dotenv>=0.21.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.3; platform_system != 'Windows'
psycopg2>=2.9.3; platform_system == 'Windows'
pymysql>=1.0.2
openai>=1.0.0
langchain>=0.0.200
langchain-community>=0.0.10
langchain-openai>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers
nltk